Changed to a new theme and added a new post about a past project.
I wasn't entirely happy with the Jekyll theme I was using, so I switched
to the theme I am now using with this commit, Slate+Simple.

I also made a few modifications to the theme, which I will made in
a fork of the git repo and thus talk about theme there.

While I was at it, (and testing out the projects section I added)
I went ahead an added a bare bones project post about the sunflower sop.
I decided to date the project with today's (2019-10-21) date since I am
posting it today, but in reality the last commit made to that repo
was December of last year, and the last edit to the guide itself was
done in 2016...